requirements.txt
agno[all]
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
groq
ddgs
fastapi
faiss-cpu
sentence-transformers
plotly